:truck: Move layout folder
#75 레이아웃 폴더를 src 하위 폴더로 위치 변경
