+++ src/Layout/NoHeaderLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Footer from '../../components/Footer/Footer';
+import Footer from '../components/Footer/Footer';
 
 const NoHeaderLayout = () => (
     <>
