+++ src/Layout/defaultLayout.jsx
@@ -1,5 +1,5 @@
-import Header from '../../components/Header/Header';
-import Footer from '../../components/Footer/Footer';
+import Header from '../components/Header/Header';
+import Footer from '../components/Footer/Footer';
 import { Outlet } from 'react-router-dom';
 const DefaultLayout = () => (
     <>
